C4 · El custodio entra en la plantilla del inquilino
`malla/41-el-cofre.yaml` — Deployment, Service, cuenta y sus dos politicas de
red, y `gen-inquilino.py` lo emite como un fichero mas del inquilino. Con la
imagen `cofre` en el Dockerfile y en `cloudbuild.yaml`, y desplegada por CI.

⛔⛔ Y AQUI SE PIERDE LA GARANTIA DE `scratch` OTRA VEZ, dicho en el Dockerfile.
Igual que `ore-serve` la perdio para tener `git`, este la pierde para tener
`gcloud` — que es lo unico con lo que habla con el KMS. Lo que se compra
cediendola: ni TLS, ni OAuth, ni criptografia en Rust. Y lo que NO se pierde, que
es lo que la hace aceptable: su cuenta de Google alcanza UNA clave, su usuario de
base de datos no alcanza `iam` mas alla de lo justo, y su NetworkPolicy solo le
abre el DNS, el servidor de metadatos, la base y Google. Tres cerraduras sobre la
misma puerta, y ninguna es la imagen.

⭐ UNA CUENTA DE GOOGLE POR INQUILINO, y el generador la sustituye. Es lo que la
medida no habia visto: `ore-driver@` es UNA cuenta compartida por los `driver` de
todos, asi que con ese patron dar permiso sobre la llave de uno lo habria dado
sobre la de todos. Aqui `ore-cofre-<inquilino>@`, con permiso sobre una clave.

⚠️ Y el comprobador del generador salto con «de-MO-strado». Un aviso que se
dispara por una palabra de la prosa enseña a ignorarlo, y entonces deja de avisar
del `t-demo` que si importa. Ahora busca el nombre como PALABRA —lindes «no una
letra»— asi que `t-demo`, `cq-demo`, `cofre-demo@` y `tenant: demo` se siguen
cazando: todos los sitios donde el nombre es un VALOR y no una silaba.

⛔ La cuenta `cofre` y su politica se van del `99` al `41`. Alli quedaban como
copia, y un objeto declarado en dos ficheros es un objeto que diverge: el dia que
uno cambie, cual gana lo decide quien aplico el ultimo. El `99` se queda con lo
suyo — el Job que demuestra el cerrojo.

⚠️ Y la entrada al custodio, por ahora, solo desde su propio namespace. La consola
tendra que llegar, y eso es la misma pregunta que la E6 de la `0022` —la entrada
por inquilino— sin contestar. Mejor decirlo que dejar una puerta abierta «para
cuando haga falta».

Y un arreglo del C3 que CI cazo: `max(version)` devuelve `integer` y el codigo lo
leia como `i64`. Es un error de tipo en tiempo de ejecucion, no de compilacion —
la clase de fallo que solo aparece con la base delante. Con `::bigint` en el SQL,
y de paso las dos comprobaciones que fallaban sin decir por que ahora enseñan el
cuerpo de la respuesta.

+++ crates/ore-cofre/src/rutas.rs
@@ -217,7 +217,11 @@ impl Servidor {
             potestad::exige(tx, emisor, &s.persona, &org, "secreto:listar")?;
             let filas = tx.filas(
                 "select s.nombre, s.clase, s.en, (s.retirado_en is not null) as retirado,
-                        coalesce(max(m.version), 0) as versiones
+                        -- ⛔ `::bigint` a proposito: `version` es `integer`, y `max`
+                        --   devuelve `integer`. Leerlo como `i64` desde el driver es un
+                        --   error de tipo en tiempo de ejecucion, no de compilacion —
+                        --   la clase de fallo que solo aparece con la base delante.
+                        coalesce(max(m.version), 0)::bigint as versiones
                    from cofre.secreto s
                    left join cofre.material m on m.secreto = s.id
                   where s.organizacion = $1
